Introduce configurable discovery algorithm
The `EngineDiscoveryRequestResolver` class implements a test discovery
in a way that can be reused by different test engines. The Jupiter and
Vintage test engines are changed to use it.

Main entry point is `EngineDiscoveryRequestResolver.builder()` which
allows to add resolvers (via a new `SelectorResolvers` interface) and
visitors (reusing `TestDescriptor.Visitor`) that will be called to
resolve `DiscoverySelectors` in a given `EngineDiscoveryRequest`.

In order to support class based test engines, the predefined
`ClassContainerSelectorResolver` may be added via the builder as well.
It resolves `ClasspathRootSelectors`, `ModuleSelectors`, and
`PackageSelectors` into `ClassSelectors` by scanning for classes in the
corresponding class container.

The fundamental principle behind `SelectorResolver` is that both the
designated parent of a `TestDescriptor` and its designated children can
be specified using `DiscoverySelectors`. For example, in the case of the
Jupiter engine, when resolving a nested test class, there's a resolver
that specifies the parent using the `ClassSelector` of the enclosing
class and the children as the set of `MethodSelectors` of the testable
methods it contains joined with the set of `ClassSelectors` of nested
classes it contains.

Overall, the algorithm works as follows:

1. Enqueue all selectors in the supplied request to be resolved.
2. While there are selectors to be resolved, get the next one.
   Otherwise, the resolution is finished.
   a. Iterate over all registered `SelectorResolvers` in the order they
      were registered in and find the first one that returns a non-empty
      resolution.
   b. If such a resolution exists, enqueue the contained selectors.
   c. For each exact match in the resolution, expand its children and
      enqueue them as well.
   d. Iterate over all registered visitors and let the engine descriptor
      accept them.

Descriptors with the same unique ID are only resolved once, even if
specified using multiple selectors.

Resolves #1739.